fdt: obey CONFIG_CMDLINE_EXTEND
We rely on CONFIG_CMDLINE_EXTEND: we want to be able to specify default
kernel parameters built into the kernel, and then have the firmware
add further parameters on top when passed at boot (currently used in
verbose boot mode for enabling serial console).

Before this patch, booting in verbose mode was wiping out the parameters
specified inside the kernel, resulting in a galcore-related crash on boot
due to not having its parameter set.

Signed-off-by: Daniel Drake <dsd@laptop.org>
[dilinger@queued.net: fix missing braces with the non-extendingohgodthebabyiscryingagaincode]