imx-drm: hdmi: get rid of pointless spinlock
The irq_lock spinlock is doing nothing for us: imx_hdmi_fb_registered()
is called from the probe function, and even though this disables local
interrupts, it doesn't stop the interrupt being serviced by another CPU
in a SMP system.

Moreover, holding a spinlock over clk_prepare*() is a bug in itself.  So
let's just get rid of this.

While we're here, ->fb_reg does nothing for us either: it's only tested
on entry to this function and set on successful exit from it, yet this
function won't be called a second time.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>