net: fec: Don't pulldown phy_reset_gpio during reset
This causes the ar8035 to not reset properly.  This is probably
due to the fact that the default state for the reset is pulled
down.

Regardless this is how upstream kernels reset the phy so I assume
that it is working for all the hardware.  This will probably need
to be revisited and fixed properly.